Deshechos varios cambios en los routers
